% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_certara.R
\name{certara_pal}
\alias{certara_pal}
\title{Certara color palette}
\usage{
certara_pal(choose)
}
\arguments{
\item{choose}{The indices of colors requested in the palette (optional). If
missing, all available colors will be included. The palette currently
contains 12 distinct colors; if more colors are needed for a plot these 12
colors will be recycled.}
}
\value{
A palette function, that returns a vector of hex codes for the
colors in the palette when called with a single integer argument.
}
\description{
Certara color palette
}
\examples{
certara_pal()(8)
}
