% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\docType{data}
\name{default.labels}
\alias{default.labels}
\title{Default labels}
\format{
An object of class \code{list} of length 7.
}
\usage{
default.labels
}
\description{
Default labels
}
\keyword{datasets}
