% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gof.R
\name{gof_list}
\alias{gof_list}
\alias{gof_layout}
\alias{gof}
\title{List of gof plots}
\usage{
gof_list(
  data = NULL,
  labels = getOption("gof.labels"),
  baseplot = gof_baseplot,
  rundir = getwd(),
  ...
)

gof_layout(p, layout = c(ceiling(length(p)/2), 2))

gof(
  data = NULL,
  panels = c(3, 4, 7, 8, 9, 10),
  layout = c(ceiling(length(panels)/2), 2),
  labels = getOption("gof.labels"),
  baseplot = gof_baseplot,
  rundir = getwd(),
  ...
)
}
\arguments{
\item{data}{A \code{data.frame}.}

\item{labels}{A named \code{list} of labels.}

\item{baseplot}{A function that returns a \code{ggplot} object to use as the
base for all scatterplots (histograms and QQ-plots are unaffected).}

\item{rundir}{A directory in which to search.}

\item{...}{Additional arguments, passed to \code{baseplot}.}

\item{p}{A \code{list} of `ggplot` objects to be laid out.}

\item{layout}{A \code{numeric} vector of length 2 giving the number of rows
and column in which the panels are to be layed out (for multiple panels).}

\item{panels}{A \code{numeric} vector specifying the panels desired.}
}
\description{
\code{gof} generats a set of panels of goodness-of-fit diagnostic plots.
}
\details{
The default plot contains the following 6 panels in a 3x2 matrix:
\itemize{
  \item \code{\link{gof_dv_vs_ipred}}
  \item \code{\link{gof_dv_vs_pred}}
  \item \code{\link{gof_cwres_vs_pred}}
  \item \code{\link{gof_cwres_vs_time}}
  \item \code{\link{gof_cwres_vs_tad}}
  \item \code{\link{gof_absiwres_vs_ipred}}
}
}
