% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gof.R
\name{gof_baseplot}
\alias{gof_baseplot}
\title{Base plot for gof plots}
\usage{
gof_baseplot(data)
}
\arguments{
\item{data}{A \code{data.frame}.}
}
\value{
A \code{ggplot} object.
}
\description{
This is the default base plot for goodness-of-fit diagnostic plots. It
produces a scatterplot with a LOESS smoother. The points are blue, and the
LOESS smoother is red. The plot has a square aspect ratio. It uses the
Certara theme. Other elements (e.g. reference lines) can be added to the
plot afterwards. The default aesthetics (x and y) and labels are also added
subsequently. The base plot can be overridden using the \code{baseplot}
parameter of the specific GOF function.
}
