% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gof.R
\name{gof_default_panels}
\alias{gof_default_panels}
\title{The default GOF panels}
\usage{
gof_default_panels(...)
}
\arguments{
\item{...}{Ignored.}
}
\value{
A vector of panel numbers.
}
\description{
The default GOF panels
}
\details{
The default output of \code{\link{gof}} is consists of 6 panels arranged in
a 3-by-2 grid. These are:

\enumerate{
  \item dv_vs_ipred
  \item dv_vs_pred
  \item cwres_vs_pred
  \item cwres_vs_time
  \item cwres_vs_tad
  \item absiwres_vs_ipred
}

Most panels come a both a linear and log variant. Normally, the linear
variant will be chosen, but in some cases it may make more sense for the log
variants to be the default for certain panels. This can be achieved using
\code{\link{options}} \code{gof.scale.*} (see Examples).
}
\examples{
gof_default_panels()

\dontrun{
# Change the default scale to log for dv/pred/ipred
options(gof.scale.dv="log")
gof_default_panels()

# Change the default scale to log for time and tad
options(gof.scale.time="log", gof.scale.tad="log")
gof_default_panels()
}
}
