% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gof.R
\name{gof_identity}
\alias{gof_identity}
\title{A generic function for identity plots}
\usage{
gof_identity(
  data,
  x,
  y,
  labels = default.labels,
  baseplot = gof_baseplot,
  log_xy = F,
  ...
)
}
\arguments{
\item{data}{A \code{data.frame}.}

\item{x, y}{Numeric vectors, evaulated within \code{data}.}

\item{labels}{A named \code{list} of labels.}

\item{baseplot}{A function that returns a \code{ggplot} object to use as the
base.}

\item{log_xy}{If \code{TRUE} then log-scale will be used for both x- and y-axes.}

\item{...}{Additional arguments, passed to \code{baseplot}.}
}
\value{
A \code{ggplot} object.
}
\description{
Identity plots have symmetric x- and y-axes, a reference line at $y=x$, and
a fixed aspect ration. Typical examples are DV vs. PRED and DV vs. IPRED.
}
