% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{gof_labels}
\alias{gof_labels}
\title{Labels for GOF plots}
\usage{
gof_labels(...)
}
\arguments{
\item{...}{Named character arguments will override the defaults.}
}
\value{
A named list.
}
\description{
Labels for GOF plots
}
\details{
This function can be used to get and set the axis labels for GOF
plots.  The default labels are taken from \code{\link{options}} that start
with the prefix \code{gof.label.}, for example \code{gof.label.dv},
\code{gof.label.ipred} and so on.  This function extracts those options into
a named list (without the prefix), and allows any number of them to be
overridden.
}
\examples{
gof_labels()

# Override a label
gof_labels(tad="Time After Start of Infusion")

\dontrun{
# Change the default
options(gof.label.tad="Time After Start of Infusion")
gof_labels()
}
}
