% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gof.R
\name{gof_qqplot}
\alias{gof_qqplot}
\title{A generic function for QQ-plots}
\usage{
gof_qqplot(data, x, labels = gof_labels())
}
\arguments{
\item{data}{A \code{data.frame}.}

\item{x}{A numeric vector, evaulated within \code{data}.}

\item{labels}{A named \code{list} of labels.}
}
\value{
A \code{ggplot} object.
}
\description{
A generic function for QQ-plots
}
