% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gof.R
\name{gof_read_data}
\alias{gof_read_data}
\title{Read gof data}
\usage{
gof_read_data(rundir = getwd())
}
\arguments{
\item{rundir}{A directory in which to search.}
}
\value{
A \code{data.frame}.
}
\description{
Read gof data
}
