% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gof.R
\name{gof_residual}
\alias{gof_residual}
\alias{gof_absresidual}
\title{A generic function for residual plots}
\usage{
gof_residual(
  data,
  x,
  y,
  labels = getOption("gof.labels"),
  baseplot = gof_baseplot,
  log_x = F,
  ...
)

gof_absresidual(
  data,
  x,
  y,
  labels = getOption("gof.labels"),
  baseplot = gof_baseplot,
  log_x = F,
  ...
)
}
\arguments{
\item{data}{A \code{data.frame}.}

\item{x, y}{Numeric vectors, evaulated within \code{data}.}

\item{labels}{A named \code{list} of labels.}

\item{baseplot}{A function that returns a \code{ggplot} object to use as the
base.}

\item{log_x}{If \code{TRUE} then log-scale will be used for the x-axis.}

\item{...}{Additional arguments, passed to \code{baseplot}.}
}
\value{
A \code{ggplot} object.
}
\description{
A generic function for residual plots
}
