% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gof.R
\name{gofplots}
\alias{gofplots}
\alias{gof_cwres_vs_pred}
\alias{gof_cwres_vs_time}
\alias{gof_cwres_vs_tad}
\alias{gof_absiwres_vs_ipred}
\alias{gof_absiwres_vs_time}
\alias{gof_absiwres_vs_tad}
\alias{gof_dv_vs_ipred}
\alias{gof_dv_vs_pred}
\alias{gof_cwres_histogram}
\alias{gof_cwres_qqplot}
\title{Basic gof plots}
\usage{
gof_cwres_vs_pred(
  data,
  labels = getOption("gof.labels"),
  baseplot = gof_baseplot,
  log_x = F,
  ...
)

gof_cwres_vs_time(
  data,
  labels = getOption("gof.labels"),
  baseplot = gof_baseplot,
  log_x = F,
  ...
)

gof_cwres_vs_tad(
  data,
  labels = getOption("gof.labels"),
  baseplot = gof_baseplot,
  log_x = F,
  ...
)

gof_absiwres_vs_ipred(
  data,
  labels = getOption("gof.labels"),
  baseplot = gof_baseplot,
  log_x = F,
  ...
)

gof_absiwres_vs_time(
  data,
  labels = getOption("gof.labels"),
  baseplot = gof_baseplot,
  log_x = F,
  ...
)

gof_absiwres_vs_tad(
  data,
  labels = getOption("gof.labels"),
  baseplot = gof_baseplot,
  log_x = F,
  ...
)

gof_dv_vs_ipred(
  data,
  labels = getOption("gof.labels"),
  baseplot = gof_baseplot,
  log_xy = F,
  ...
)

gof_dv_vs_pred(
  data,
  labels = getOption("gof.labels"),
  baseplot = gof_baseplot,
  log_xy = F,
  ...
)

gof_cwres_histogram(data, labels = getOption("gof.labels"))

gof_cwres_qqplot(data, labels = getOption("gof.labels"))
}
\arguments{
\item{data}{A \code{data.frame}.}

\item{labels}{A named \code{list} of labels.}

\item{baseplot}{A function that returns a \code{ggplot} object to use as the
base for all scatterplots (histograms and QQ-plots are unaffected).}

\item{log_x}{If \code{TRUE} then log-scale will be used for the x-axis.}

\item{...}{Additional arguments, passed to \code{baseplot}.}

\item{log_xy}{If \code{TRUE} then log-scale will be used for both x- and y-axes.}
}
\value{
A \code{ggplot} object.
}
\description{
Basic gof plots

Plot CWRES vs. PRED

Plot CWRES vs. TIME

Plot CWRES vs. TAD

Plot |IWRES| vs. IPRED

Plot |IWRES| vs. TIME

Plot |IWRES| vs. TAD

Plot DV vs. IPRED

Plot DV vs. PRED

Histogram of CWRES

QQ-plot of CWRES
}
