% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gof.R
\name{indiv_plot}
\alias{indiv_plot}
\title{Generate Individual Plots}
\usage{
indiv_plot(
  data,
  x,
  y,
  ipred,
  pred,
  labels = gof_labels(),
  ip_ncol = 4,
  ip_nrow = 4,
  scale_facet = "fixed",
  labelfacet = "label_value",
  strati1 = id,
  strati2 = NULL,
  color_color = c(dv = ggcertara::certara_pal()[1], outliers =
    ggcertara::certara_pal()[2], ipred = ggcertara::certara_pal()[4], pred =
    ggcertara::certara_pal()[7]),
  fill_fill = c(ipred = ggcertara::certara_pal()[4]),
  alpha_fill = 0.5,
  line_linetype = c(dv = "blank", outliers = "blank", ipred = "solid", pred = "dashed"),
  line_size = 1,
  shape_shape = c(dv = 16, outliers = 10, ipred = NA, pred = NA),
  shape_size = 1,
  cwres.outliers = FALSE,
  limit.outliers = 5,
  uncertainty = FALSE,
  uncert.ci = 0.95,
  ...
)
}
\arguments{
\item{data}{A \code{data.frame} containing the dataset.}

\item{x}{The name of the independent variable column. Default value is "time".}

\item{y}{The name of the dependent variable column. Default value is "dv".}

\item{ipred}{The name of the individual prediction variable column. Default value is "ipred".}

\item{pred}{The name of the population prediction variable column. Default value is "pred".}

\item{labels}{A named \code{list} of labels for plot annotations. Default values are provided by \code{gof_labels()}.}

\item{ip_ncol}{Number of columns per plot. Default is 4. Adjust if encountering issues with incomplete first rows on the last page due to \code{ggforce::facet_wrap_paginate}.}

\item{ip_nrow}{Number of rows per plot. Default is 4. Adjust similarly to \code{ip_ncol} for pagination issues.}

\item{strati1}{The variable used to stratify the plots. Default is "id".}

\item{strati2}{An optional variable for an additional layer of stratification. Default is NULL.}

\item{color_color}{A named vector specifying colors for different elements (dv, ipred, pred, outliers). Should be a vector of 4 colors.}

\item{fill_fill}{A named vector specifying fill colors for elements when uncertainty is plotted. Specifically for 'ipred' when \code{uncertainty} is TRUE.}

\item{alpha_fill}{Transparency level for the filled area representing the confidence interval, when \code{uncertainty} is TRUE.}

\item{line_linetype}{A named vector specifying line types for different elements (dv, outliers, ipred, pred).}

\item{line_size}{Line width for the line layer. Default is 1.}

\item{shape_shape}{A named vector specifying shapes for different elements. When not specified, defaults are used.}

\item{shape_size}{Size for the shape layer. Default is 1.}

\item{cwres.outliers}{If \code{TRUE}, only profiles of individuals with outlier observations are displayed. Default is FALSE.}

\item{limit.outliers}{Threshold for CWRES to identify outliers. Default is 5.}

\item{uncertainty}{If \code{TRUE}, plots a confidence interval around individual predictions. Default is FALSE.}

\item{uncert.ci}{Confidence interval level when \code{uncertainty} is TRUE. Default is 0.95.}

\item{...}{Additional arguments passed to lower-level functions.}
}
\description{
Observations (DV), individual predictions (IPRED), and population predictions (PRED) plotted against the independent variable for each individual.
}
\examples{
\dontrun{
indiv_plot(data = dat,
           x = "time",
           y = "dv",
           ipred = "ipred",
           pred = "pred",
           strati1 = "id",
           ip_ncol = 3,
           ip_nrow = 4,
           uncertainty = FALSE)
}

# Example with individuals with high cwres values
\dontrun{
indiv_plot(data = dat,
           x = "time",
           y = "dv",
           ipred = "ipred",
           pred = "pred",
           strati1 = "id",
           cwres.outliers = TRUE,
           limit.outliers = 5,
           ip_ncol = 3,
           ip_nrow = 4,
           uncertainty = FALSE)
}

# Example to apply layers with the ampersand sign & (example to change legend.position)
\dontrun{
indiv_plot(data = dat,
           x = "time",
           y = "dv",
           ipred = "ipred",
           pred = "pred",
           strati1 = "id") &
theme(legend.position = "bottom")
}
}
