% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gof.R
\name{logscales}
\alias{logscales}
\alias{log_x}
\alias{log_y}
\alias{log_xy}
\title{Use log scales}
\usage{
log_x(g)

log_y(g)

log_xy(g)
}
\arguments{
\item{g}{A \code{ggplot2} object.}
}
\value{
g A \code{ggplot2} object with log scales in $x$, $y$ or both.
}
\description{
Use log scales
}
