% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_colorexplorer.R
\name{run_colorexplorer}
\alias{run_colorexplorer}
\title{Run the colorexplorer shiny app}
\usage{
run_colorexplorer(plotobj = NULL)
}
\arguments{
\item{plotobj}{You can pass a \code{ggplot} object to be used in the shiny
app. It should have at least either a \code{color} or \code{fill} aesthetic
for this to make sense (otherwise, you won't be able to change any of the
colors interactively).
#' @return Called for its side effects.}
}
\description{
Run the colorexplorer shiny app
}
\section{Note}{

The app requires the following additional packages:
\itemize{
  \item `shiny`
  \item `colourpicker`
  \item `colorblindr` (available on GitHub at \url{https://github.com/clauswilke/colorblindr})
}
Make sure they are installed or the app won't work.
}

\examples{
\dontrun{
library(ggplot2)

g <- ggplot(mpg, aes(class, fill=class)) + geom_bar()
run_colorexplorer(g)
}
}
