% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_certara.R
\name{theme_certara}
\alias{theme_certara}
\alias{theme_certara_grid}
\alias{scale_colour_certara}
\alias{scale_color_certara}
\alias{scale_fill_certara}
\alias{scale_colour_certara_c}
\alias{scale_fill_certara_c}
\title{A ggplot2 theme for Certara CSC-iDD}
\usage{
theme_certara(base_size = 16, base_family = "",
  base_line_size = base_size/22, base_rect_size = base_size/22)

theme_certara_grid(base_size = 16, base_family = "",
  base_line_size = base_size/22, base_rect_size = base_size/22)

scale_colour_certara(...)

scale_color_certara(...)

scale_fill_certara(...)

scale_colour_certara_c(..., guide = "colourbar")

scale_fill_certara_c(..., guide = "colourbar")
}
\arguments{
\item{base_size}{base font size}

\item{base_family}{base font family}

\item{base_line_size}{base size for line elements}

\item{base_rect_size}{base size for rect elements}

\item{...}{Other arguments passed on to [discrete_scale()] or
[continuous_scale()]}

\item{guide}{A function used to create a guide or its name. See
\code{\link[=guides]{guides()}} for more info.}
}
\value{
An object of class \code{\link[ggplot2]{theme}()}.
}
\description{
A ggplot2 theme for Certara CSC-iDD
}
\examples{

\donttest{
library(ggplot2)

set.seed(123)
x <- rnorm(100)
g <- rep(0:5, len=100)
y <- rnorm(100, 0.33*x + 0.2*g, 0.1)
g <- factor(g, labels=c("xxx", "yyy", "zzz", "aaa", "bbb", "ccc"))

theme_set(theme_certara())
scale_colour_discrete <- function(...) scale_color_certara(...)
scale_fill_discrete <- function(...) scale_fill_certara(...)

gl <- guide_legend(title.position="top")
p <- ggplot(data.frame(x, y, g), aes(x=x, y=y, color=g, shape=g, linetype=g, fill=g)) +
    geom_point() +
    geom_smooth(method="lm", se=T) +
    labs(title="Random Data", subtitle="test-test") +
    guides(colour=gl, fill=gl)
p

p + theme_certara_grid()
p + scale_y_continuous(position="right")

p + facet_wrap(~ g)

v <- ggplot(faithfuld) + geom_tile(aes(waiting, eruptions, fill=density))
v + scale_fill_certara_c()
}
}
