% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_certara.R
\name{certara_pal}
\alias{certara_pal}
\title{Certara color palette}
\usage{
certara_pal(n = 17)
}
\arguments{
\item{n}{The number of colors requested in the palette.}
}
\value{
A character vector of hex codes for the colors in the palette.
}
\description{
Certara color palette
}
\examples{
certara_pal(8)
}
