% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_certara.R
\name{certara_pal}
\alias{certara_pal}
\title{Certara color palettes}
\usage{
certara_pal(palnum = 1)
}
\arguments{
\item{palnum}{Choose from 1 of 5 different color palettes with 8 colors each.}
}
\value{
A palette function, that returns a vector of hex codes for the
colors in the palette when called with a single integer argument.
}
\description{
Certara color palettes
}
\examples{
certara_pal(1)(8)
}
