% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gof.R
\name{gof_list}
\alias{gof_list}
\alias{gof_layout}
\alias{gof}
\title{List of gof plots}
\usage{
gof_list(
  data = NULL,
  panels = gof_default_panels(),
  empty = FALSE,
  all = FALSE,
  labels = gof_labels(),
  baseplot = gof_baseplot,
  rundir = getwd(),
  ...
)

gof_layout(p, layout = NULL, transpose = FALSE, ...)

gof(
  data = NULL,
  panels = gof_default_panels(),
  layout = c(ceiling(length(panels)/2), 2),
  transpose = FALSE,
  labels = gof_labels(),
  baseplot = gof_baseplot,
  rundir = getwd(),
  ...
)
}
\arguments{
\item{data}{A \code{data.frame}.}

\item{panels}{A \code{numeric} vector specifying the panels desired.}

\item{labels}{A named \code{list} of labels.}

\item{baseplot}{A function that returns a \code{ggplot} object to use as the
base for all scatterplots (histograms and QQ-plots are unaffected).}

\item{rundir}{A directory in which to search.}

\item{...}{Additional arguments passed to other methods (e.g. \code{\link{baseplot}}).}

\item{p}{A \code{list} of `ggplot` objects to be laid out.}

\item{layout}{A \code{numeric} vector of length 2 giving the number of rows
and column in which the panels are to be layed out (for multiple panels).}

\item{transpose}{Should the layout be transposed (rows become columns)?}
}
\description{
List of gof plots

Layout gof plots

Goodness-of-fit diagnostic plots
}
