% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gof.R
\name{gof_list}
\alias{gof_list}
\alias{gof_layout}
\alias{gof}
\title{List of gof plots}
\usage{
gof_list(
  data = NULL,
  panels = gof_default_panels(),
  empty = FALSE,
  all = FALSE,
  labels = gof_labels(),
  baseplot = gof_baseplot,
  rundir = getwd(),
  ...
)

gof_layout(p, nrow = NULL, ncol = NULL, byrow = TRUE, ...)

gof(
  data = NULL,
  panels = gof_default_panels(),
  layout = NULL,
  labels = gof_labels(),
  baseplot = gof_baseplot,
  rundir = getwd(),
  ...
)
}
\arguments{
\item{data}{A \code{data.frame}.}

\item{panels}{A \code{numeric} vector specifying the panels desired.}

\item{empty}{Return an empty list.}

\item{all}{Return a list of all panels.}

\item{labels}{A named \code{list} of labels.}

\item{baseplot}{A function that returns a \code{ggplot} object to use as the
base for all scatterplots (histograms and QQ-plots are unaffected).}

\item{rundir}{A directory in which to search.}

\item{...}{Additional arguments passed to other methods (e.g. \code{baseplot}).}

\item{p}{A \code{list} of `ggplot` objects to be laid out.}

\item{ncol, nrow}{The dimensions of the grid to create - if both are \code{NULL} it
will use the same logic as \link[ggplot2:facet_wrap]{facet_wrap()} to set the
dimensions}

\item{byrow}{Analogous to \code{byrow} in \link[base:matrix]{matrix()}. If \code{FALSE} the
plots will be filled in in column-major order}

\item{layout}{A \code{list} of arguments to pass to \code{\link{gof_layout}} (for multiple panels).}
}
\description{
List of gof plots

Layout gof plots

Goodness-of-fit diagnostic plots
}
\seealso{
\code{\link[patchwork]{plot_layout}}
}
