% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gof.R
\name{gof_baseplot}
\alias{gof_baseplot}
\title{Base plot for gof plots}
\usage{
gof_baseplot(data, highlight, ..., loess.args = list())
}
\arguments{
\item{data}{A \code{data.frame}.}

\item{highlight}{(Optional) A 2-level \code{factor}, evaluated within
\code{data}, indicating a subset of points that are to be highlighted on the
plot.  Typically, these may be outliers, or a specific subset of interest.
Those points belonging to the subset will be drawn with a different color
and symbol, and a legend will appear as well.}

\item{...}{Additional arguments (ignored).}
}
\value{
A \code{ggplot} object.
}
\description{
This is the default base plot for goodness-of-fit diagnostic plots. It
produces a scatterplot with a LOESS smoother. The points are blue, and the
LOESS smoother is red. The plot has a square aspect ratio. It uses the
Certara theme. Other elements (e.g. reference lines) can be added to the
plot afterwards. The default aesthetics (x and y) and labels are also added
subsequently. The base plot can be overridden using the \code{baseplot}
parameter of the specific GOF function.
}
