% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gof.R
\name{gof_read_data}
\alias{gof_read_data}
\title{Read gof data}
\usage{
gof_read_data(file = NULL, rundir = getwd())
}
\arguments{
\item{file}{A filename to read the data from. If not specified, a simple
heuristic search in \code{rundir} is performed.}

\item{rundir}{A directory in which to search.}
}
\value{
A \code{data.frame}.
}
\description{
This function reads in data files with the .csv file extension and files containing
the string "sdtab" from the current working directory. Names are set to lower case
and data items are checked to verify that "dv", "pred", and "ipred" are present.
}
