% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gof.R
\name{indiv_plot}
\alias{indiv_plot}
\title{A generic function for individual plots.}
\usage{
indiv_plot(
  data,
  x,
  y,
  ipred,
  pred,
  labels = gof_labels(),
  ip_ncol = 4,
  ip_nrow = 4,
  scale_facet = "fixed",
  labelfacet = "label_value",
  strati1 = id,
  strati2 = NULL,
  color_color = c(dv = ggcertara::certara_pal()[1], outliers =
    ggcertara::certara_pal()[2], ipred = ggcertara::certara_pal()[4], pred =
    ggcertara::certara_pal()[7]),
  fill_fill = c(ipred = ggcertara::certara_pal()[4]),
  alpha_fill = 0.5,
  line_linetype = c(dv = "blank", outliers = "blank", ipred = "solid", pred = "dashed"),
  line_size = 1,
  shape_shape = c(dv = 16, outliers = 10, ipred = NA, pred = NA),
  shape_size = 1,
  cwres.outliers = FALSE,
  limit.outliers = 5,
  uncertainty = FALSE,
  uncert.ci = 0.95,
  ...
)
}
\arguments{
\item{data}{A \code{data.frame}.}

\item{x}{Changes independent variable. Value is time by default}

\item{y}{Changes dependent variable. Value is dv by default}

\item{ipred}{Changes individual prediction variable. Value is ipred by default}

\item{pred}{Changes population prediction variable. Value is pred by default}

\item{labels}{A named \code{list} of labels.}

\item{ip_ncol}{Variable used to define numbers of columns per plot. Value is 4 by default. Works in combination with ip_nrow.
A bug occurs when the first row of last page not complete (due to ggforce::facet_wrap_paginate). In this case you get a error message asking you to change ip_ncol and/or ip_nrow}

\item{ip_nrow}{Variable used to define numbers of rows per plot. Value is 4 by default. Works in combination with ip_ncol.
A bug occurs when the first row of last page not complete (due to ggforce::facet_wrap_paginate). In this case you get a error message asking you to change ip_ncol and/or ip_nrow}

\item{strati1}{Variable used to stratify the plots. Value is id by default.}

\item{strati2}{Variable used to possibly add an additional layer of stratification to the plots. Value is NULL by default.}

\item{color_color}{Changes color. Should be a vector of 4 values (i.e. dv, ipred, pred, outliers).}

\item{fill_fill}{Changes fill color for ipred when plotting with uncertainty option set to TRIE.}

\item{alpha_fill}{Transparency parameter for the area plotted to represent confidence interval.}

\item{line_linetype}{Changes linetype.}

\item{line_size}{Size parameter for the line layer.}

\item{shape_shape}{Changes shape.}

\item{shape_size}{Size parameter for the shape layer.}

\item{cwres.outliers}{If \code{TRUE} then only profiles of individuals with outliers observations will be displayed. Value is FALSE by default.}

\item{limit.outliers}{Value to be used for CWRES to identify outliers. Value is 5 by default.}

\item{uncertainty}{If \code{TRUE} then  confidence interval (95% by default) around individual predictions are plotted with observations.}

\item{uncert.ci}{Value used to define the confidence interval when having the uncertainty flag set to TRUE () 0.95 by default.}
}
\description{
Observations (DV), individual predictions (IPRED) and population predictions (PRED) plotted against the independent variable for each individual.
}
\examples{
#Examples
\dontrun{
indiv_plot(data = dat,
           strati1 = id,
           x = time,
           y = dv,
           ipred = ipred,
           pred = pred,
           ip_nrow = 4,
           ip_ncol = 3,
           uncertainty = FALSE)
}

# Example with individuals with high cwres values
\dontrun{
indiv_plot(data = dat,
           strati1 = id,
           x = time,
           y = dv,
           ipred = ipred,
           pred = pred,
           ip_nrow = 4,
           ip_ncol = 3,
           cwres.outliers = TRUE,
           limit.outliers = 5,
           uncertainty = FALSE)
}

# Example to apply layers with the ampersand sign & (example to change legend.position)
\dontrun{
indiv_plot(data = dat,
           strati1 = id,
           x = time,
           y = dv,
           ipred = ipred,
           pred = pred) &
theme(legend.position = "bottom")
}
}
