% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gof.R
\name{logscales}
\alias{logscales}
\alias{log_x}
\alias{log_y}
\alias{log_xy}
\title{Use log scales}
\usage{
log_x(g, limits = NULL)

log_y(g, limits = NULL)

log_xy(g, limits = NULL)
}
\arguments{
\item{g}{A \code{ggplot2} object.}

\item{limits}{Limits to use for the scale. See \code{\link[ggplot2]{scale_x_continuous}}.}
}
\value{
g A \code{ggplot2} object with log scales in \eqn{x}, \eqn{y} or both.
}
\description{
This function puts the respective axes in log scale.
}
