% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_certara.R
\name{theme_certara}
\alias{theme_certara}
\alias{theme_certara_grid}
\alias{theme_certara_hgrid}
\alias{scale_colour_certara}
\alias{scale_color_certara}
\alias{scale_fill_certara}
\alias{scale_colour_certara_c}
\alias{scale_fill_certara_c}
\title{A ggplot2 theme for Certara CSC-iDD}
\usage{
theme_certara(
  base_size = 11,
  base_family = "",
  base_line_size = base_size/22,
  base_rect_size = base_size/22,
  grid = c("none", "horizontal", "both")
)

theme_certara_grid(
  base_size = 11,
  base_family = "",
  base_line_size = base_size/22,
  base_rect_size = base_size/22
)

theme_certara_hgrid(
  base_size = 11,
  base_family = "",
  base_line_size = base_size/22,
  base_rect_size = base_size/22
)

scale_colour_certara(...)

scale_color_certara(...)

scale_fill_certara(...)

scale_colour_certara_c(..., guide = "colourbar")

scale_fill_certara_c(..., guide = "colourbar")
}
\arguments{
\item{base_size}{base font size, given in pts.}

\item{base_family}{base font family}

\item{base_line_size}{base size for line elements}

\item{base_rect_size}{base size for rect elements}

\item{...}{Other arguments passed on to \code{\link[ggplot2]{discrete_scale}()} or
\code{\link[ggplot2]{continuous_scale}()}.}

\item{guide}{A function used to create a guide or its name. See
\code{\link[ggplot2:guides]{guides()}} for more information.}
}
\value{
An object of class \code{\link[ggplot2]{theme}()}.
}
\description{
A ggplot2 theme for Certara CSC-iDD
}
\details{
There are 3 variants of the theme: no grid
(\code{theme_certara()}), full grid (\code{theme_certara_grid()}), and
horizontal grid lines only (\code{theme_certara_hgrid()}).
}
\examples{

\donttest{
library(ggplot2)

set.seed(123)
x <- rnorm(100) + 3
g <- rep(0:5, len=100)
y <- exp(rnorm(100, 0.33*x + 0.2*g, 0.1))
g <- factor(g, labels=do.call(paste0, rep(list(letters[1:6]), 3)))

theme_set(theme_certara())
scale_colour_discrete <- function(...) scale_color_certara(...)
scale_fill_discrete <- function(...) scale_fill_certara(...)

gl <- guide_legend("Legend", title.position="top")
p <- ggplot(data.frame(x, y, g), aes(x=x, y=y, color=g, shape=g, linetype=g, fill=g)) +
    geom_point() +
    geom_smooth(method="lm", se=T) +
    labs(title="Random Data", subtitle="test-test") +
    guides(colour=gl, fill=gl, shape=gl, linetype=gl)
p

p + theme_certara_grid()
p + scale_y_continuous(position="right")

p + facet_wrap(~ g)
p + scale_y_log10() + annotation_logticks(sides="l")
p + scale_y_log10() + annotation_logticks(sides="l") + theme_certara_grid()

v <- ggplot(faithfuld) + geom_tile(aes(waiting, eruptions, fill=density))
v + scale_fill_certara_c()

p <- function(n=8) {
    dat <- data.frame(x=1:n)
    ggplot(dat) + geom_col(aes(x=x, y=1, fill=factor(x)), color=NA, show.legend=F)
}
p(8) + scale_fill_certara()
p(8) + scale_fill_certara(start=11)
p(4) + scale_fill_certara(select=c(14, 19, 1, 5))
p(8) + scale_fill_certara(start=9, discard=4)
p(8) + scale_fill_certara(permute=c(2, 4, 1))
p(8) + scale_fill_certara(permute=c(2, 4, 1), discard=3)
p(30) + scale_fill_certara()

}
}
